chore: split tui::state into code+mod declarations
